+++ src/raw/client.rs
@@ -130,14 +130,6 @@ impl Client {
         self.post(Endpoint::DeleteListen, token, data)
     }
 
-    /// Endpoint: [`users/{user_list}/recent-listens`](https://listenbrainz.readthedocs.io/en/production/dev/api/#get--1-users-(user_list)-recent-listens)
-    pub fn users_recent_listens(
-        &self,
-        user_list: &[&str],
-    ) -> Result<UsersRecentListensResponse, Error> {
-        self.get(Endpoint::UsersRecentListens(user_list))
-    }
-
     /// Endpoint: [`user/{user_name}/listen-count`](https://listenbrainz.readthedocs.io/en/production/dev/api/#get--1-user-(user_name)-listen-count)
     pub fn user_listen_count(&self, user_name: &str) -> Result<UserListenCountResponse, Error> {
         self.get(Endpoint::UserListenCount(user_name))
@@ -287,7 +279,7 @@ impl Client {
         user_name: &str,
         range: Option<&str>,
         force_recalculate: Option<bool>,
-    ) -> Result<StatsUserArtistMapResponse, Error> {
+    ) -> Result<Option<StatsUserArtistMapResponse>, Error> {
         let endpoint = format!(
             "{}{}",
             API_ROOT_URL,
@@ -305,7 +297,12 @@ impl Client {
 
         let response = request.send()?;
 
-        ResponseType::from_response(response)
+        // API returns 204 and an empty document if there are no statistics
+        if response.status() == 204 {
+            Ok(None)
+        } else {
+            ResponseType::from_response(response).map(Some)
+        }
     }
 
     /// Endpoint: [`stats/user/{user_name}/releases`](https://listenbrainz.readthedocs.io/en/production/dev/api/#get--1-stats-user-(user_name)-releases)